Integrated UCT with TicTacToe game
If you just run main.py, it'll be human vs model.
However, the model does not seem to be training correctly, as it picks
pretty terrible moves. In fact, it seems to be intentionally picking
the worst moves, so I suspect that perhaps the values are flipped,
but changing the points_values and win_values does not seem to do
anything.